requirements.txt
exif>=1.6.1
httpx>=0.28.1
pydantic>=2.12.0
tqdm>=4.67.1
pillow
